updates wip

+++ install/run_app_backend.sh
@@ -1,7 +1,12 @@
 #!/bin/sh
 workdir=$(pwd)
-apt-get update
-apt-get install -y \
+setup_done=false
+
+if [ -f /.setup_done ] ;then
+  echo "setup done"
+else
+  apt-get update
+  apt-get install -y \
     git \
     gcc \
     g++ \
@@ -9,18 +14,10 @@ apt-get install -y \
     openexr \
     && rm -rf /var/lib/apt/lists/*
 
-pip install --upgrade pip
-
-if [ -f /app/src/launcher.py ]; then
-  pip install --no-cache-dir -r /app/requirements.txt
-  echo "Files exist. Please remove the old files/ update to install new one."
-else
-  rm -rf $/app/
-  mkdir -p /app
-  cd /app
-
+  pip install --upgrade pip
+  setup_done=true
+  touch /.setup_done
   git clone https://github.com/BreakTools/deadline-web-app-backend .
-  pip install --no-cache-dir -r /app/requirements.txt
+  pip install -r /app/requirements.txt
 fi
-
 python /app/src/launcher.py
